+++ src/Stack/Options/ConfigParser.hs
@@ -64,11 +64,11 @@ configOptsParser currentDir hide0 =
     <*> buildOptsMonoidParser hide0
     <*> dockerOptsParser True
     <*> nixOptsParser True
-    <*> firstBoolFlags
+    <*> firstBoolFlagsNoDefault
             "system-ghc"
             "using the system installed GHC (on the PATH) if it is available and its version matches. Disabled by default."
             hide
-    <*> firstBoolFlags
+    <*> firstBoolFlagsTrue
             "install-ghc"
             "downloading and installing GHC if necessary (can be done manually with stack setup)"
             hide
@@ -113,11 +113,11 @@ configOptsParser currentDir hide0 =
             <> help "Use HPACK executable (overrides bundled Hpack)"
             <> hide
              ))
-    <*> firstBoolFlags
+    <*> firstBoolFlagsFalse
             "skip-ghc-check"
             "skipping the GHC version and architecture check"
             hide
-    <*> firstBoolFlags
+    <*> firstBoolFlagsFalse
             "skip-msys"
             "skipping the local MSYS installation (Windows only)"
             hide
@@ -128,19 +128,20 @@ configOptsParser currentDir hide0 =
             <> help "Install binaries to DIR"
             <> hide
              ))
-    <*> firstBoolFlags
+    <*> firstBoolFlagsTrue
             "modify-code-page"
             "setting the codepage to support UTF-8 (Windows only)"
             hide
-    <*> firstBoolFlags
+    <*> firstBoolFlagsNoDefault
             "allow-different-user"
             ("permission for users other than the owner of the stack root " ++
-                "directory to use a stack installation (POSIX only)")
+                "directory to use a stack installation (POSIX only) " ++
+                "(default: true inside Docker, otherwise false)")
             hide
     <*> fmap toDumpLogs
-            (firstBoolFlags
+            (firstBoolFlagsNoDefault
              "dump-logs"
-             "dump the build output logs for local packages to the console"
+             "dump the build output logs for local packages to the console (default: dump warning logs)"
              hide)
     <*> optionalFirst (option readColorWhen
              ( long "color"